feat: :zap: finish instructions and vm almost finished

+++ corewar/include/instruction.h
@@ -50,6 +50,6 @@ static instruction_t *const all_instruction[] = {
 };
 
     #define NB_INSTRUCTION sizeof(all_instruction) / sizeof(instruction_t *)
-    #define IS_INSTRUCTION(nb) (0 < nb && (int)nb < (int)NB_INSTRUCTION + 1)
+    #define IS_INSTRUCTION(nb) (0 < nb && (unsigned int)nb < NB_INSTRUCTION + 1)
 
 #endif /* !INSTRUCTION_H_ */
